fix: logout
dispatch logout throught wrapper
